Solve exercise 8
One function was created for subtraction, multiplication, power, and
factorial. For power, a for loop was used, meanwhile for factorial the
function was recursive. Two functions were created for addition, the
first admits two arguments, the second admits just one but it has to be
an array. For the latter, the reduce function for arrays was used