Solve exercise 12
The function creates a new array with object elements that have as
properties the names and the ages of each pearson. Then, it sorts the
array based on age from oldest to youngest, and returns the object from
the original array that has the same name as the first element of the
new array. If the original element doesn't have a year of death, it uses
the current year to calculate the age of the pearson